Fix "signed and unsigned type in conditional expression" warning

+++ examples/alffplay.c
@@ -686,7 +686,7 @@ static int audio_thread(void *userdata)
         movState->audio.st->codec->sample_rate,
         movState->audio.st->codec->channel_layout ?
             movState->audio.st->codec->channel_layout :
-            av_get_default_channel_layout(movState->audio.st->codec->channels),
+            (uint64_t)av_get_default_channel_layout(movState->audio.st->codec->channels),
         movState->audio.st->codec->sample_fmt,
         movState->audio.st->codec->sample_rate,
         0, NULL
